1.补充删除时二次确认；

+++ src/page/projectList.vue
@@ -74,7 +74,7 @@
         ></el-pagination>
       </div>
 
-      <el-dialog title="编辑" :visible.sync="editDialogFormVisible">
+      <el-dialog title="编辑" :visible.sync="editDialogFormVisible"  :close-on-click-modal=false>
         <el-form :model="dataInfo">
           <el-form-item label="*项目名称" label-width="120px">
             <el-input v-model="dataInfo.name"  size='small'></el-input>
@@ -104,7 +104,7 @@
         </div>
       </el-dialog>
 
-      <el-dialog title="添加" :visible.sync="addDialogFormVisible">
+      <el-dialog title="添加" :visible.sync="addDialogFormVisible" :close-on-click-modal=false>
         <el-form :model="dataAdd" ref="dataAdd">
           <el-form-item label="*项目名称" label-width="120px" prop="name">
             <el-input v-model="dataAdd.name" auto-complete="off"  size='small'></el-input>
@@ -208,22 +208,28 @@ export default {
       }
     },
     async handleDelete(projectId, index) {
-      const res = await deleteProject(projectId);
-      if (res.code == 200) {
-        this.$message({
-          type: "success",
-          center: true,
-          message: res.msg
-        });
-        this.total --;
-        this.dataList.splice(index, 1);
-      } else {
-        this.$message({
-          type: "error",
-          center: true,
-          message: res.msg
-        });
-      }
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(async () => {
+        const res = await deleteProject(projectId);
+        if (res.code == 200) {
+          this.$message({
+            type: "success",
+            center: true,
+            message: res.msg
+          });
+          this.total --;
+          this.dataList.splice(index, 1);
+        } else {
+          this.$message({
+            type: "error",
+            center: true,
+            message: res.msg
+          });
+        }
+      })  
     },
     async updateProject() {
       const res = await modifyProject(this.dataInfo);
