+++ src/page/userList.vue
@@ -66,7 +66,7 @@
         ></el-pagination>
       </div>
 
-      <el-dialog title="编辑" :visible.sync="editDialogFormVisible">
+      <el-dialog title="编辑" :visible.sync="editDialogFormVisible" :close-on-click-modal=false>
         <el-form :model="userInfo">
           <el-form-item label="*用户名" label-width="100px">
             <el-input v-model="userInfo.username" auto-complete="off" size='small'></el-input>
@@ -99,7 +99,7 @@
         </div>
       </el-dialog>
 
-      <el-dialog title="新增" :visible.sync="addDialogFormVisible">
+      <el-dialog title="新增" :visible.sync="addDialogFormVisible" :close-on-click-modal=false>
         <el-form :model="addForm" ref="addForm">
           <el-form-item label="*用户名" label-width="100px" prop="username">
             <el-input v-model="addForm.username" auto-complete="off" size='small'></el-input>
@@ -208,23 +208,29 @@ export default {
         });
       }
     },
-    async handleDelete(userId, index) {
-      const res = await deleteUser(userId);
-      if (res.code == 200) {
-        this.$message({
-          type: "success",
-          center: true,
-          message: res.msg
-        });
-        this.total --;
-        this.tableData.splice(index, 1);
-      } else {
-        this.$message({
-          type: "error",
-          center: true,
-          message: res.msg
-        });
-      }
+    handleDelete(userId, index) {
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(async () => {
+        let res = await deleteUser(userId);
+        if (res.code == 200) {
+          this.$message({
+            type: "success",
+            center: true,
+            message: res.msg
+          });
+          this.total --;
+          this.tableData.splice(index, 1);
+        } else {
+          this.$message({
+            type: "error",
+            center: true,
+            message: res.msg
+          });
+        }
+      })
     },
     async updateUser() {
       const res = await modifyUserInfo(this.userInfo);
